Extracted shared code from scenes
