+++ js/scenes/lobby-scene.js
@@ -1,35 +1,14 @@
 import { Chicken } from "../entities/chicken.js";
 import { CustomizeOverlay } from "../ui/customize-overlay.js";
-import { NetworkSync } from "../engine/network-sync.js";
 import { RunnerScene } from "./runner-scene.js";
+import { BaseScene } from "./base-scene.js";
 
-export class LobbyScene {
-  /**
-   * @param {{ctx: CanvasRenderingContext2D, canvas: HTMLCanvasElement, viewport: { width: number, height: number }, assets: Object, input: import('../engine/input.js').InputManager, network: import('../engine/network.js').NetworkManager, switchScene: Function, cloudLayer: Object, rng: Object}} context
-   */
+export class LobbyScene extends BaseScene {
   constructor(context) {
-    this.ctx = context.ctx;
-    this.canvas = context.canvas;
-    this.viewport = context.viewport;
-    this.canvasW = this.viewport.width;
-    this.canvasH = this.viewport.height;
-    this.assets = context.assets;
-    this.input = context.input;
-    this.network = context.network;
-    this.switchScene = context.switchScene;
-    this.cloudLayer = context.cloudLayer;
-    this.rng = context.rng;
-    this.horizonY = context.horizonY;
-    this.context = context;
-    this.chicken = null;
+    super(context);
     this.overlay = null;
-    this.networkSync = null;
-    this.cameraX = 0;
-
     this.localReady = false;
-    this.readyPlayers = new Set(); // ids of ready remote players
-
-    this.onKeyDown = null;
+    this.readyPlayers = new Set();
   }
 
   enter() {
@@ -39,7 +18,7 @@ export class LobbyScene {
     });
 
     // random spawn
-    const pad = 144; // 3 * shape of the chicken sprite
+    const pad = 144;
     this.chicken.x = pad + Math.random() * (this.canvasW - this.chicken.width - pad * 2);
     this.chicken.y = this.chicken.minY + Math.random() * (this.chicken.maxY - this.chicken.minY);
     this.chicken.facingRight = Math.random() < 0.5;
@@ -60,13 +39,12 @@ export class LobbyScene {
     );
 
     // network sync
-    this.networkSync = new NetworkSync(this.network, this.assets);
-    this.networkSync.init(this.chicken, appearance);
+    this.initNetworkSync(appearance);
 
     // refresh overlay button highlights when server assigns a color
     const origOnId = this.network.onId;
     this.network.onId = (colorIndex) => {
-      origOnId(colorIndex);
+      origOnId?.(colorIndex);
       this.overlay.syncStyleSelection();
     };
 
@@ -78,7 +56,6 @@ export class LobbyScene {
     this.readyPlayers.clear();
 
     this.network.onReady = (id, ready) => {
-      // ignore our own ready echo — tracked by localReady
       if (id === this.network.id) return;
       if (ready) {
         this.readyPlayers.add(id);
@@ -91,11 +68,9 @@ export class LobbyScene {
       this.startRunner(roundSeed);
     };
 
-    // reset ready states when someone joins/leaves
     const origOnJoin = this.network.onJoin;
     this.network.onJoin = (id, colorIndex, spriteSet, name) => {
       origOnJoin?.(id, colorIndex, spriteSet, name);
-      // new player joined — reset ready states
       this.localReady = false;
       this.readyPlayers.clear();
     };
@@ -107,15 +82,7 @@ export class LobbyScene {
       this.localReady = false;
     };
 
-    // ground tile config
-    this.tileSize = 16;
-    this.tileScale = 3;
-    this.drawSize = this.tileSize * this.tileScale;
-    this.groundRows = (this.canvasH - this.horizonY) / this.drawSize;
-
-    const { groundTileset, groundEdgeTileset } = this.assets.environment;
-    this.edgeTileCount = groundEdgeTileset.width / this.tileSize;
-    this.groundTileCount = groundTileset.width / this.tileSize;
+    this.initGround();
 
     // listen for Enter to toggle ready / start in single player
     this.onKeyDown = (e) => {
@@ -124,7 +91,6 @@ export class LobbyScene {
           this.network.sendReady();
           this.localReady = !this.localReady;
         } else {
-          // offline single-player: generate a random seed locally
           const roundSeed = Math.floor(Math.random() * 0x7fffffff);
           this.startRunner(roundSeed);
         }
@@ -139,7 +105,6 @@ export class LobbyScene {
     this.switchScene(runner);
   }
 
-  /** @param {number} dt */
   update(dt) {
     this.cloudLayer.update();
     this.chicken.update(dt);
@@ -150,30 +115,18 @@ export class LobbyScene {
     if (this.chicken.x < 0) this.chicken.x = 0;
     if (this.chicken.x > maxX) this.chicken.x = maxX;
 
-    // keep camera fixed in lobby
     this.cameraX = 0;
   }
 
   render() {
-    const ctx = this.ctx;
-
-    ctx.clearRect(0, 0, this.canvasW, this.canvasH);
-
-    this.cloudLayer.render(ctx, this.cameraX);
-
-    this.renderGround(ctx);
+    this.renderBackground();
 
     // draw all chickens sorted by y
     const allChickens = [this.chicken, ...this.networkSync.getRemoteChickens()];
-    allChickens.sort((a, b) => a.y - b.y);
-    for (const chicken of allChickens) {
-      const origX = chicken.x;
-      chicken.x = origX - this.cameraX;
-      chicken.render(ctx);
-      chicken.x = origX;
-    }
+    this.renderChickens(allChickens);
 
-    // draw ready hint at the bottom
+    // draw ready hint
+    const ctx = this.ctx;
     ctx.save();
     const remoteCount = this.networkSync.getRemoteChickens().length;
     const totalPlayers = 1 + remoteCount;
@@ -191,56 +144,13 @@ export class LobbyScene {
     ctx.restore();
   }
 
-  renderGround(ctx) {
-    const horizonY = this.horizonY;
-    const { tileSize, drawSize, groundRows } = this;
-    const { groundTileset, groundEdgeTileset } = this.assets.environment;
-
-    const startCol = Math.floor(this.cameraX / drawSize);
-    const visibleCols = Math.ceil(this.canvasW / drawSize) + 1;
-
-    ctx.save();
-    ctx.imageSmoothingEnabled = false;
-    for (let row = 0; row < groundRows; row++) {
-      const isEdge = row === 0;
-      const tileset = isEdge ? groundEdgeTileset : groundTileset;
-      const tileCount = isEdge ? this.edgeTileCount : this.groundTileCount;
-      for (let i = 0; i < visibleCols; i++) {
-        const worldCol = startCol + i;
-        const hash = this.rng.hash(worldCol, row);
-        const tileIdx = hash % tileCount;
-        const screenX = worldCol * drawSize - this.cameraX;
-        ctx.drawImage(
-          tileset,
-          tileIdx * tileSize,
-          0,
-          tileSize,
-          tileSize,
-          Math.round(screenX),
-          horizonY + row * drawSize,
-          drawSize,
-          drawSize,
-        );
-      }
-    }
-    ctx.restore();
-  }
-
   exit() {
-    if (this.onKeyDown) {
-      window.removeEventListener("keydown", this.onKeyDown);
-      this.onKeyDown = null;
-    }
     this.network.onReady = null;
     this.network.onStart = null;
-    this.chicken = null;
-    if (this.networkSync) {
-      this.networkSync.destroy();
-      this.networkSync = null;
-    }
     if (this.overlay) {
       this.overlay.destroy();
       this.overlay = null;
     }
+    super.exit();
   }
 }
